+++ src/main/java/me/ajaja/module/tag/application/CreatePlanTagService.java
@@ -11,8 +11,8 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import me.ajaja.module.tag.adapter.out.persistence.PlanTagRepository;
 import me.ajaja.module.tag.adapter.out.persistence.TagRepository;
-import me.ajaja.module.tag.adapter.out.persistence.model.PlanTag;
-import me.ajaja.module.tag.adapter.out.persistence.model.Tag;
+import me.ajaja.module.tag.domain.PlanTag;
+import me.ajaja.module.tag.domain.Tag;
 import me.ajaja.module.tag.application.port.out.CreateTagPort;
 
 @Service
