+++ src/main/java/me/ajaja/module/tag/domain/PlanTag.java
@@ -1,13 +1,11 @@
-package me.ajaja.module.tag.adapter.out.persistence.model;
+package me.ajaja.module.tag.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
+import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,21 +13,22 @@ import me.ajaja.global.common.BaseEntity;
 
 @Getter
 @Entity
-@Table(name = "tags")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Tag extends BaseEntity<Tag> {
+public class PlanTag extends BaseEntity<PlanTag> {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "tag_id")
+	@Column(name = "plan_tag_id")
 	private Long id;
 
-	@NotBlank
-	@Size(max = 10)
-	@Column(name = "tag_name", unique = true)
-	private String name;
+	@NotNull
+	private Long planId;
 
-	public Tag(String name) {
-		this.name = name;
+	@NotNull
+	private Long tagId;
+
+	public PlanTag(Long planId, Long tagId) {
+		this.planId = planId;
+		this.tagId = tagId;
 		this.validateSelf();
 	}
 }
