+++ src/main/java/me/ajaja/module/tag/domain/Tag.java
@@ -1,11 +1,13 @@
-package me.ajaja.module.tag.adapter.out.persistence.model;
+package me.ajaja.module.tag.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.validation.constraints.NotNull;
+import jakarta.persistence.Table;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,22 +15,21 @@ import me.ajaja.global.common.BaseEntity;
 
 @Getter
 @Entity
+@Table(name = "tags")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PlanTag extends BaseEntity<PlanTag> {
+public class Tag extends BaseEntity<Tag> {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "plan_tag_id")
+	@Column(name = "tag_id")
 	private Long id;
 
-	@NotNull
-	private Long planId;
+	@NotBlank
+	@Size(max = 10)
+	@Column(name = "tag_name", unique = true)
+	private String name;
 
-	@NotNull
-	private Long tagId;
-
-	public PlanTag(Long planId, Long tagId) {
-		this.planId = planId;
-		this.tagId = tagId;
+	public Tag(String name) {
+		this.name = name;
 		this.validateSelf();
 	}
 }
